requirements.txt
auto-py-to-exe
libusb1
pip
pyaml
pyinstaller
pywin32-ctypes
setuptools
wheel
pywin32
screeninfo>=0.8.1
click==8.1.7
retry>=0.9.2
pydantic
pydantic-settings
pyyaml==6.0.1
pyusb==1.2.1
psutil==5.9.8
retrying==1.3.4